requirements.txt
sentence-transformers
faiss-cpu
pypdf
python-docx
tqdm
numpy
fastapi
uvicorn[standard]
python-multipart
pydantic>=2.0
